pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.4.0",
    "fastapi[standard]>=0.135.3",
    "jupyterlab>=4.5.6",
    "langchain>=1.2.14",
    "langchain-anthropic>=1.4.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.24",
    "langchain-huggingface>=1.0.0",
    "langchain-text-splitters>=1.0.0",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv.workspace]
members = [
    ".",
]

[tool.uv.sources]
rag-pipeline = { workspace = true, editable = true }

[dependency-groups]
dev = [
    "rag-pipeline",
]
